Add MatroskaJS Chapter Codec
Rather than reinventing the wheel with a custom language, we can reuse and well-known one.

The goal of Matroska Script was to be similar to JavaScript/ECMAStrip/ActionScript.
But, in JS, integers can only use 56 bits. So not all ChapterUID (64-bit unsigned integer) can be addressed
with the current Matroska Script stable API.

Rather than adding a different GotoAndPlay that is compatible with JavaScript (using a string), we start from
scratch with ECMAScript 6th Edition (ES6).